portalGalt/requirements.txt
asgiref==3.8.1
coverage==7.6.10
Django==5.1.4
djangorestframework==3.15.2
nose==1.3.7
numpy==2.2.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
